# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Go App Layout API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/tenant/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "Create tenant",
                "parameters": [
                    {
                        "description": "Create tenant request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.TenantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Create tenant response",
                        "schema": {
                            "$ref": "#/definitions/api.Response-model_Tenant"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Response-model_Tenant": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/model.Tenant"
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "model.Tenant": {
            "type": "object",
            "properties": {
                "api_key": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "v1.TenantRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "tenant description",
                    "type": "string"
                },
                "name": {
                    "description": "tenant name",
                    "type": "string"
                }
            }
        }
    }
}